Move database initialization to a different module
Solve flake8 issues

Change variables to snake_case
