+++ app.py
@@ -1,82 +1,38 @@
 import json
-from flask import Flask, redirect, render_template, request, session, flash
-from flaskext.mysql import MySQL
-from datetime import datetime
-
-import os
-from flask import send_from_directory, make_response
 import cryptocode
+import os
+
+from datetime import datetime
+from flask import (flash,
+                   Flask,
+                   make_response,
+                   redirect,
+                   render_template,
+                   request, session,
+                   send_from_directory,)
+from flaskext.mysql import MySQL
+
+import database
 
 cantidad_mesas = 3
 
 app = Flask(__name__)
 
 app.secret_key = '123Prueba!'
-mysql = MySQL()
 app.config['MYSQL_DATABASE_HOST'] = 'localhost'
 app.config['MYSQL_DATABASE_USER'] = 'root'
 app.config['MYSQL_DATABASE_PASSWORD'] = ''
+app.config['CARPETA'] = os.path.join('fotos')
 
 
+mysql = MySQL()
 mysql.init_app(app)
-CARPETA = os.path.join('fotos')
-app.config['CARPETA'] = CARPETA
-
 conn = mysql.connect()
 cursor = conn.cursor()
-cursor.execute("CREATE DATABASE IF NOT EXISTS `my_resto`;")
 
-cursor.execute("""CREATE TABLE IF NOT EXISTS `my_resto`.`categorias` (
-    `id_categoria` INT(10) NOT NULL AUTO_INCREMENT,
-    `categoria` VARCHAR(255) NOT NULL,
-    PRIMARY KEY (`id_categoria`))""")
-
-cursor.execute("""CREATE TABLE IF NOT EXISTS `my_resto`.`platos` (
-    `id_plato` INT(10) NOT NULL AUTO_INCREMENT,
-    `nombre` VARCHAR(255) NOT NULL ,
-    `descripcion_plato` VARCHAR(5000) NOT NULL ,
-    `precio` FLOAT NOT NULL ,
-    `foto` VARCHAR(5000) NOT NULL,
-    `id_categoria` INT(10) NOT NULL,
-    PRIMARY KEY (`id_plato`),
-    FOREIGN KEY (`id_categoria`) REFERENCES `my_resto`.`categorias`(
-    `id_categoria`));""")
-
-cursor.execute("""CREATE TABLE IF NOT EXISTS `my_resto`.`mesas` (
-    `id_mesa` INT(10) NOT NULL AUTO_INCREMENT,
-    `pedidos` JSON DEFAULT ('{ }'),
-    `hora_abre` DATETIME,
-    PRIMARY KEY (`id_mesa`));""")
-
-cursor.execute("""CREATE TABLE IF NOT EXISTS `my_resto`.`usuarios`(
-    `usuario` VARCHAR(255) NOT NULL,
-    `password` VARCHAR(500) NOT NULL,
-    `super_usuario` BOOLEAN NULL DEFAULT FALSE,
-    PRIMARY KEY (`usuario`))""")
-
-cursor.execute("""CREATE TABLE IF NOT EXISTS `my_resto`.`ventas`(
-    `id_venta` INT(20) NOT NULL AUTO_INCREMENT,
-    `mesa`INT(10),
-    `hora_abre` DATETIME ,
-    `hora_cierra` DATETIME,
-    `consumo` JSON NOT NULL DEFAULT ('{ }'),
-    `total` INT(10),
-    PRIMARY KEY (`id_venta`));""")
-
-cursor.execute("SELECT count(*) FROM `my_resto`.`usuarios`")
-cantidadDeUsuarios = cursor.fetchone()[0]
-cursor.execute("SELECT count(*) FROM `my_resto`.`categorias`")
-cantidadCategorias = cursor.fetchone()[0]
-
-if cantidadDeUsuarios == 0:
-    clave = cryptocode.encrypt('admin', app.secret_key)
-    cursor.execute("""INSERT `my_resto`.`usuarios`(
-        `usuario`,`password`,`super_usuario`)
-        VALUES ('admin', %s, 1);""", (clave))
-if cantidadCategorias == 0:
-    cursor.execute("""INSERT IGNORE `my_resto`.`categorias` (`categoria`)
-    VALUES('Sin categoria')""")
-conn.commit()
+database.create(mysql)
+database.create_admin_user(mysql, app.secret_key)
+database.define_default_category(mysql)
 
 
 @app.route('/')
